switched mapper to output JSON instead of pickle
pickle has different protocol versions. since i use 3 for most
everything but spark only understands 2, i would need to muck around
with the output to make it compatible with 2. it's just easier to use
json. it is just a dictionary after all.

also added spark version of the feature extractor. it is nearly
complete. as always, string replacement with ints is the problem child.